requirements.txt
pyperclip3
pyqrcode
string
secrets